+++ src/Task/TaskBuilder.php
@@ -11,7 +11,6 @@
 namespace Task;
 
 use Cron\CronExpression;
-use Task\Scheduler\SchedulerInterface;
 
 /**
  * Builder for tasks.
